requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
lightgbm>=3.3.0
xgboost>=1.5.0
matplotlib>=3.5.0
PyYAML>=6.0
scipy>=1.7.0
python-dateutil>=2.8.2
